Moved a bunch of files, and added readmes
