+++ app/src/main/java/fnsb/macstransit/Activities/ActivityListeners/InfoWindowAdapter.java
@@ -1,4 +1,4 @@
-package fnsb.macstransit.ActivityListeners;
+package fnsb.macstransit.Activities.ActivityListeners;
 
 import android.graphics.Color;
 import android.view.View;
@@ -7,7 +7,7 @@ import android.widget.TextView;
 
 import com.google.android.gms.maps.model.Marker;
 
-import fnsb.macstransit.MapsActivity;
+import fnsb.macstransit.Activities.MapsActivity;
 
 /**
  * Created by Spud on 2019-11-01 for the project: MACS Transit.
