+++ app/src/main/java/fnsb/macstransit/Activities/ActivityListeners/readme.md
@@ -1,0 +1,17 @@
+# Understanding
+TODO
+
+## AdjustZoom.java
+TODO
+
+## Helpers.java
+TODO
+
+## InfoWindowAdapter
+TODO
+
+## StopClicked.java
+TODO
+
+## StopDeselected.java
+TODO
