+++ app/src/main/java/fnsb/macstransit/Activities/readme.md
@@ -1,0 +1,11 @@
+# Understanding
+TODO
+
+## ActivityListeners
+TODO
+
+## MapsActivity.java
+TODO
+
+## SplashActivity.java
+TODO
