+++ app/src/main/java/fnsb/macstransit/RouteMatch/readme.md
@@ -1,6 +1,7 @@
-#Understanding
+# Understanding
+TODO
 
-## Basic Stop
+## BasicStop.java
 TODO
 
 ## Bus.java
