+++ app/src/main/java/fnsb/macstransit/Threads/readme.md
@@ -1,0 +1,8 @@
+# Understanding
+TODO
+
+## Network.java
+TODO
+
+## UpdateThread.java
+TODO
