+++ app/src/main/java/fnsb/macstransit/readme.md
@@ -4,14 +4,11 @@ TODO
 ## Flow overview
 TODO
 
-## SplashActivity.java
+## Activities
 TODO
 
-## MapsActivity.java
+## RouteMatch
 TODO
 
-## UpdateThread.java
-TODO
-
-## Network.java
+## Threads
 TODO
