+++ docs/allclasses-frame.html
@@ -12,12 +12,12 @@
 <h1 class="bar">All&nbsp;Classes</h1>
 <div class="indexContainer">
 <ul>
-<li><a href="fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.ActivityListeners" target="classFrame">AdjustZoom</a></li>
+<li><a href="fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.Activities.ActivityListeners" target="classFrame">AdjustZoom</a></li>
 <li><a href="fnsb/macstransit/RouteMatch/BasicStop.html" title="class in fnsb.macstransit.RouteMatch" target="classFrame">BasicStop</a></li>
 <li><a href="fnsb/macstransit/RouteMatch/Bus.html" title="class in fnsb.macstransit.RouteMatch" target="classFrame">Bus</a></li>
 <li><a href="fnsb/macstransit/RouteMatch/Heading.html" title="enum in fnsb.macstransit.RouteMatch" target="classFrame">Heading</a></li>
-<li><a href="fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.ActivityListeners" target="classFrame">Helpers</a></li>
-<li><a href="fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.ActivityListeners" target="classFrame">InfoWindowAdapter</a></li>
+<li><a href="fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.Activities.ActivityListeners" target="classFrame">Helpers</a></li>
+<li><a href="fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.Activities.ActivityListeners" target="classFrame">InfoWindowAdapter</a></li>
 <li><a href="fnsb/macstransit/MapsActivity.html" title="class in fnsb.macstransit" target="classFrame">MapsActivity</a></li>
 <li><a href="fnsb/macstransit/Network.html" title="class in fnsb.macstransit" target="classFrame">Network</a></li>
 <li><a href="fnsb/macstransit/RouteMatch/Route.html" title="class in fnsb.macstransit.RouteMatch" target="classFrame">Route</a></li>
@@ -25,8 +25,8 @@
 <li><a href="fnsb/macstransit/RouteMatch/SharedStop.html" title="class in fnsb.macstransit.RouteMatch" target="classFrame">SharedStop</a></li>
 <li><a href="fnsb/macstransit/SplashActivity.html" title="class in fnsb.macstransit" target="classFrame">SplashActivity</a></li>
 <li><a href="fnsb/macstransit/RouteMatch/Stop.html" title="class in fnsb.macstransit.RouteMatch" target="classFrame">Stop</a></li>
-<li><a href="fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.ActivityListeners" target="classFrame">StopClicked</a></li>
-<li><a href="fnsb/macstransit/ActivityListeners/StopDeselected.html" title="class in fnsb.macstransit.ActivityListeners" target="classFrame">StopDeselected</a></li>
+<li><a href="fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.Activities.ActivityListeners" target="classFrame">StopClicked</a></li>
+<li><a href="fnsb/macstransit/ActivityListeners/StopDeselected.html" title="class in fnsb.macstransit.Activities.ActivityListeners" target="classFrame">StopDeselected</a></li>
 <li><a href="fnsb/macstransit/UpdateThread.html" title="class in fnsb.macstransit" target="classFrame">UpdateThread</a></li>
 </ul>
 </div>
