+++ docs/allclasses-noframe.html
@@ -12,12 +12,12 @@
 <h1 class="bar">All&nbsp;Classes</h1>
 <div class="indexContainer">
 <ul>
-<li><a href="fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.ActivityListeners">AdjustZoom</a></li>
+<li><a href="fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.Activities.ActivityListeners">AdjustZoom</a></li>
 <li><a href="fnsb/macstransit/RouteMatch/BasicStop.html" title="class in fnsb.macstransit.RouteMatch">BasicStop</a></li>
 <li><a href="fnsb/macstransit/RouteMatch/Bus.html" title="class in fnsb.macstransit.RouteMatch">Bus</a></li>
 <li><a href="fnsb/macstransit/RouteMatch/Heading.html" title="enum in fnsb.macstransit.RouteMatch">Heading</a></li>
-<li><a href="fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.ActivityListeners">Helpers</a></li>
-<li><a href="fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.ActivityListeners">InfoWindowAdapter</a></li>
+<li><a href="fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.Activities.ActivityListeners">Helpers</a></li>
+<li><a href="fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.Activities.ActivityListeners">InfoWindowAdapter</a></li>
 <li><a href="fnsb/macstransit/MapsActivity.html" title="class in fnsb.macstransit">MapsActivity</a></li>
 <li><a href="fnsb/macstransit/Network.html" title="class in fnsb.macstransit">Network</a></li>
 <li><a href="fnsb/macstransit/RouteMatch/Route.html" title="class in fnsb.macstransit.RouteMatch">Route</a></li>
@@ -25,8 +25,8 @@
 <li><a href="fnsb/macstransit/RouteMatch/SharedStop.html" title="class in fnsb.macstransit.RouteMatch">SharedStop</a></li>
 <li><a href="fnsb/macstransit/SplashActivity.html" title="class in fnsb.macstransit">SplashActivity</a></li>
 <li><a href="fnsb/macstransit/RouteMatch/Stop.html" title="class in fnsb.macstransit.RouteMatch">Stop</a></li>
-<li><a href="fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.ActivityListeners">StopClicked</a></li>
-<li><a href="fnsb/macstransit/ActivityListeners/StopDeselected.html" title="class in fnsb.macstransit.ActivityListeners">StopDeselected</a></li>
+<li><a href="fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.Activities.ActivityListeners">StopClicked</a></li>
+<li><a href="fnsb/macstransit/ActivityListeners/StopDeselected.html" title="class in fnsb.macstransit.Activities.ActivityListeners">StopDeselected</a></li>
 <li><a href="fnsb/macstransit/UpdateThread.html" title="class in fnsb.macstransit">UpdateThread</a></li>
 </ul>
 </div>
