+++ docs/fnsb/macstransit/ActivityListeners/AdjustZoom.html
@@ -49,7 +49,7 @@ var activeTableTab = "activeTableTab";
 <div class="subNav">
 <ul class="navList">
 <li>Prev&nbsp;Class</li>
-<li><a href="../../../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
 <li><a href="../../../index.html?fnsb/macstransit/ActivityListeners/AdjustZoom.html" target="_top">Frames</a></li>
@@ -91,7 +91,7 @@ var activeTableTab = "activeTableTab";
 <!-- ========= END OF TOP NAVBAR ========= -->
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle">fnsb.macstransit.ActivityListeners</div>
+<div class="subTitle">fnsb.macstransit.Activities.ActivityListeners</div>
 <h2 title="Class AdjustZoom" class="title">Class AdjustZoom</h2>
 </div>
 <div class="contentContainer">
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>fnsb.macstransit.ActivityListeners.AdjustZoom</li>
+<li>fnsb.macstransit.Activities.ActivityListeners.AdjustZoom</li>
 </ul>
 </li>
 </ul>
@@ -142,7 +142,7 @@ implements com.google.android.gms.maps.GoogleMap.OnCameraIdleListener</pre>
 <th class="colOne" scope="col">Constructor and Description</th>
 </tr>
 <tr class="altColor">
-<td class="colOne"><code><span class="memberNameLink"><a href="../../../fnsb/macstransit/ActivityListeners/AdjustZoom.html#AdjustZoom-fnsb.macstransit.MapsActivity-">AdjustZoom</a></span>(<a href="../../../fnsb/macstransit/MapsActivity.html" title="class in fnsb.macstransit">MapsActivity</a>&nbsp;activity)</code>
+<td class="colOne"><code><span class="memberNameLink"><a href="../../../fnsb/macstransit/ActivityListeners/AdjustZoom.html#AdjustZoom-fnsb.macstransit.Activities.MapsActivity-">AdjustZoom</a></span>(<a href="../../../fnsb/macstransit/MapsActivity.html" title="class in fnsb.macstransit">MapsActivity</a>&nbsp;activity)</code>
 <div class="block">Constructor for the listener.</div>
 </td>
 </tr>
@@ -197,7 +197,7 @@ implements com.google.android.gms.maps.GoogleMap.OnCameraIdleListener</pre>
 <!--   -->
 </a>
 <h3>Constructor Detail</h3>
-<a name="AdjustZoom-fnsb.macstransit.MapsActivity-">
+<a name="AdjustZoom-fnsb.macstransit.Activities.MapsActivity-">
 <!--   -->
 </a>
 <ul class="blockListLast">
@@ -282,7 +282,7 @@ implements com.google.android.gms.maps.GoogleMap.OnCameraIdleListener</pre>
 <div class="subNav">
 <ul class="navList">
 <li>Prev&nbsp;Class</li>
-<li><a href="../../../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
 <li><a href="../../../index.html?fnsb/macstransit/ActivityListeners/AdjustZoom.html" target="_top">Frames</a></li>
