+++ docs/fnsb/macstransit/ActivityListeners/Helpers.html
@@ -48,8 +48,8 @@ var activeTableTab = "activeTableTab";
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../../fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
-<li><a href="../../../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../../fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
 <li><a href="../../../index.html?fnsb/macstransit/ActivityListeners/Helpers.html" target="_top">Frames</a></li>
@@ -91,7 +91,7 @@ var activeTableTab = "activeTableTab";
 <!-- ========= END OF TOP NAVBAR ========= -->
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle">fnsb.macstransit.ActivityListeners</div>
+<div class="subTitle">fnsb.macstransit.Activities.ActivityListeners</div>
 <h2 title="Class Helpers" class="title">Class Helpers</h2>
 </div>
 <div class="contentContainer">
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>fnsb.macstransit.ActivityListeners.Helpers</li>
+<li>fnsb.macstransit.Activities.ActivityListeners.Helpers</li>
 </ul>
 </li>
 </ul>
@@ -344,8 +344,8 @@ extends java.lang.Object</pre>
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../../fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
-<li><a href="../../../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../../fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
 <li><a href="../../../index.html?fnsb/macstransit/ActivityListeners/Helpers.html" target="_top">Frames</a></li>
