+++ docs/fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html
@@ -48,8 +48,8 @@ var activeTableTab = "activeTableTab";
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
-<li><a href="../../../fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../../fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
 <li><a href="../../../index.html?fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" target="_top">Frames</a></li>
@@ -91,7 +91,7 @@ var activeTableTab = "activeTableTab";
 <!-- ========= END OF TOP NAVBAR ========= -->
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle">fnsb.macstransit.ActivityListeners</div>
+<div class="subTitle">fnsb.macstransit.Activities.ActivityListeners</div>
 <h2 title="Class InfoWindowAdapter" class="title">Class InfoWindowAdapter</h2>
 </div>
 <div class="contentContainer">
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>fnsb.macstransit.ActivityListeners.InfoWindowAdapter</li>
+<li>fnsb.macstransit.Activities.ActivityListeners.InfoWindowAdapter</li>
 </ul>
 </li>
 </ul>
@@ -142,7 +142,7 @@ implements com.google.android.gms.maps.GoogleMap.InfoWindowAdapter</pre>
 <th class="colOne" scope="col">Constructor and Description</th>
 </tr>
 <tr class="altColor">
-<td class="colOne"><code><span class="memberNameLink"><a href="../../../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html#InfoWindowAdapter-fnsb.macstransit.MapsActivity-">InfoWindowAdapter</a></span>(<a href="../../../fnsb/macstransit/MapsActivity.html" title="class in fnsb.macstransit">MapsActivity</a>&nbsp;activity)</code>
+<td class="colOne"><code><span class="memberNameLink"><a href="../../../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html#InfoWindowAdapter-fnsb.macstransit.Activities.MapsActivity-">InfoWindowAdapter</a></span>(<a href="../../../fnsb/macstransit/MapsActivity.html" title="class in fnsb.macstransit">MapsActivity</a>&nbsp;activity)</code>
 <div class="block">Constructor for the info window adapter.</div>
 </td>
 </tr>
@@ -195,7 +195,7 @@ implements com.google.android.gms.maps.GoogleMap.InfoWindowAdapter</pre>
 <!--   -->
 </a>
 <h3>Constructor Detail</h3>
-<a name="InfoWindowAdapter-fnsb.macstransit.MapsActivity-">
+<a name="InfoWindowAdapter-fnsb.macstransit.Activities.MapsActivity-">
 <!--   -->
 </a>
 <ul class="blockListLast">
@@ -298,8 +298,8 @@ implements com.google.android.gms.maps.GoogleMap.InfoWindowAdapter</pre>
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
-<li><a href="../../../fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
+<li><a href="../../../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../../fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">Next&nbsp;Class</span></a></li>
 </ul>
 <ul class="navList">
 <li><a href="../../../index.html?fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" target="_top">Frames</a></li>
