+++ docs/fnsb/macstransit/ActivityListeners/StopDeselected.html
@@ -48,7 +48,7 @@ var activeTableTab = "activeTableTab";
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../../fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../../fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
 <li>Next&nbsp;Class</li>
 </ul>
 <ul class="navList">
@@ -91,7 +91,7 @@ var activeTableTab = "activeTableTab";
 <!-- ========= END OF TOP NAVBAR ========= -->
 <!-- ======== START OF CLASS DATA ======== -->
 <div class="header">
-<div class="subTitle">fnsb.macstransit.ActivityListeners</div>
+<div class="subTitle">fnsb.macstransit.Activities.ActivityListeners</div>
 <h2 title="Class StopDeselected" class="title">Class StopDeselected</h2>
 </div>
 <div class="contentContainer">
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>fnsb.macstransit.ActivityListeners.StopDeselected</li>
+<li>fnsb.macstransit.Activities.ActivityListeners.StopDeselected</li>
 </ul>
 </li>
 </ul>
@@ -249,7 +249,7 @@ implements com.google.android.gms.maps.GoogleMap.OnInfoWindowCloseListener</pre>
 </div>
 <div class="subNav">
 <ul class="navList">
-<li><a href="../../../fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
+<li><a href="../../../fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">Prev&nbsp;Class</span></a></li>
 <li>Next&nbsp;Class</li>
 </ul>
 <ul class="navList">
