+++ docs/fnsb/macstransit/ActivityListeners/package-frame.html
@@ -3,21 +3,21 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (1.8.0_202-release) on Fri Nov 15 21:22:49 AKST 2019 -->
-<title>fnsb.macstransit.ActivityListeners</title>
+<title>fnsb.macstransit.Activities.ActivityListeners</title>
 <meta name="date" content="2019-11-15">
 <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../script.js"></script>
 </head>
 <body>
-<h1 class="bar"><a href="../../../fnsb/macstransit/ActivityListeners/package-summary.html" target="classFrame">fnsb.macstransit.ActivityListeners</a></h1>
+<h1 class="bar"><a href="../../../fnsb/macstransit/ActivityListeners/package-summary.html" target="classFrame">fnsb.macstransit.Activities.ActivityListeners</a></h1>
 <div class="indexContainer">
 <h2 title="Classes">Classes</h2>
 <ul title="Classes">
-<li><a href="AdjustZoom.html" title="class in fnsb.macstransit.ActivityListeners" target="classFrame">AdjustZoom</a></li>
-<li><a href="Helpers.html" title="class in fnsb.macstransit.ActivityListeners" target="classFrame">Helpers</a></li>
-<li><a href="InfoWindowAdapter.html" title="class in fnsb.macstransit.ActivityListeners" target="classFrame">InfoWindowAdapter</a></li>
-<li><a href="StopClicked.html" title="class in fnsb.macstransit.ActivityListeners" target="classFrame">StopClicked</a></li>
-<li><a href="StopDeselected.html" title="class in fnsb.macstransit.ActivityListeners" target="classFrame">StopDeselected</a></li>
+<li><a href="AdjustZoom.html" title="class in fnsb.macstransit.Activities.ActivityListeners" target="classFrame">AdjustZoom</a></li>
+<li><a href="Helpers.html" title="class in fnsb.macstransit.Activities.ActivityListeners" target="classFrame">Helpers</a></li>
+<li><a href="InfoWindowAdapter.html" title="class in fnsb.macstransit.Activities.ActivityListeners" target="classFrame">InfoWindowAdapter</a></li>
+<li><a href="StopClicked.html" title="class in fnsb.macstransit.Activities.ActivityListeners" target="classFrame">StopClicked</a></li>
+<li><a href="StopDeselected.html" title="class in fnsb.macstransit.Activities.ActivityListeners" target="classFrame">StopDeselected</a></li>
 </ul>
 </div>
 </body>
