+++ docs/fnsb/macstransit/ActivityListeners/package-summary.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (1.8.0_202-release) on Fri Nov 15 21:22:49 AKST 2019 -->
-<title>fnsb.macstransit.ActivityListeners</title>
+<title>fnsb.macstransit.Activities.ActivityListeners</title>
 <meta name="date" content="2019-11-15">
 <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../script.js"></script>
@@ -12,7 +12,7 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="fnsb.macstransit.ActivityListeners";
+            parent.document.title="fnsb.macstransit.Activities.ActivityListeners";
         }
     }
     catch(err) {
@@ -69,7 +69,7 @@
 </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-<h1 title="Package" class="title">Package&nbsp;fnsb.macstransit.ActivityListeners</h1>
+<h1 title="Package" class="title">Package&nbsp;fnsb.macstransit.Activities.ActivityListeners</h1>
 </div>
 <div class="contentContainer">
 <ul class="blockList">
@@ -82,31 +82,31 @@
 </tr>
 <tbody>
 <tr class="altColor">
-<td class="colFirst"><a href="../../../fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.ActivityListeners">AdjustZoom</a></td>
+<td class="colFirst"><a href="../../../fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.Activities.ActivityListeners">AdjustZoom</a></td>
 <td class="colLast">
 <div class="block">Created by Spud on 2019-10-28 for the project: MACS Transit.</div>
 </td>
 </tr>
 <tr class="rowColor">
-<td class="colFirst"><a href="../../../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.ActivityListeners">Helpers</a></td>
+<td class="colFirst"><a href="../../../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.Activities.ActivityListeners">Helpers</a></td>
 <td class="colLast">
 <div class="block">Created by Spud on 2019-11-01 for the project: MACS Transit.</div>
 </td>
 </tr>
 <tr class="altColor">
-<td class="colFirst"><a href="../../../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.ActivityListeners">InfoWindowAdapter</a></td>
+<td class="colFirst"><a href="../../../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.Activities.ActivityListeners">InfoWindowAdapter</a></td>
 <td class="colLast">
 <div class="block">Created by Spud on 2019-11-01 for the project: MACS Transit.</div>
 </td>
 </tr>
 <tr class="rowColor">
-<td class="colFirst"><a href="../../../fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.ActivityListeners">StopClicked</a></td>
+<td class="colFirst"><a href="../../../fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.Activities.ActivityListeners">StopClicked</a></td>
 <td class="colLast">
 <div class="block">Created by Spud on 2019-10-30 for the project: MACS Transit.</div>
 </td>
 </tr>
 <tr class="altColor">
-<td class="colFirst"><a href="../../../fnsb/macstransit/ActivityListeners/StopDeselected.html" title="class in fnsb.macstransit.ActivityListeners">StopDeselected</a></td>
+<td class="colFirst"><a href="../../../fnsb/macstransit/ActivityListeners/StopDeselected.html" title="class in fnsb.macstransit.Activities.ActivityListeners">StopDeselected</a></td>
 <td class="colLast">
 <div class="block">Created by Spud on 2019-11-11 for the project: MACS Transit.</div>
 </td>
