+++ docs/fnsb/macstransit/ActivityListeners/package-tree.html
@@ -3,7 +3,7 @@
 <html lang="en">
 <head>
 <!-- Generated by javadoc (1.8.0_202-release) on Fri Nov 15 21:22:49 AKST 2019 -->
-<title>fnsb.macstransit.ActivityListeners Class Hierarchy</title>
+<title>fnsb.macstransit.Activities.ActivityListeners Class Hierarchy</title>
 <meta name="date" content="2019-11-15">
 <link rel="stylesheet" type="text/css" href="../../../stylesheet.css" title="Style">
 <script type="text/javascript" src="../../../script.js"></script>
@@ -12,13 +12,12 @@
 <script type="text/javascript"><!--
     try {
         if (location.href.indexOf('is-external=true') == -1) {
-            parent.document.title="fnsb.macstransit.ActivityListeners Class Hierarchy";
+            parent.document.title="fnsb.macstransit.Activities.ActivityListeners Class Hierarchy";
         }
     }
     catch(err) {
     }
-//-->
-</script>
+//--></script>
 <noscript>
 <div>JavaScript is disabled on your browser.</div>
 </noscript>
@@ -69,7 +68,7 @@
 </a></div>
 <!-- ========= END OF TOP NAVBAR ========= -->
 <div class="header">
-<h1 class="title">Hierarchy For Package fnsb.macstransit.ActivityListeners</h1>
+<h1 class="title">Hierarchy For Package fnsb.macstransit.Activities.ActivityListeners</h1>
 <span class="packageHierarchyLabel">Package Hierarchies:</span>
 <ul class="horizontal">
 <li><a href="../../../overview-tree.html">All Packages</a></li>
@@ -80,11 +79,11 @@
 <ul>
 <li type="circle">java.lang.Object
 <ul>
-<li type="circle">fnsb.macstransit.ActivityListeners.<a href="../../../fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">AdjustZoom</span></a> (implements com.google.android.gms.maps.GoogleMap.OnCameraIdleListener)</li>
-<li type="circle">fnsb.macstransit.ActivityListeners.<a href="../../../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">Helpers</span></a></li>
-<li type="circle">fnsb.macstransit.ActivityListeners.<a href="../../../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">InfoWindowAdapter</span></a> (implements com.google.android.gms.maps.GoogleMap.InfoWindowAdapter)</li>
-<li type="circle">fnsb.macstransit.ActivityListeners.<a href="../../../fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">StopClicked</span></a> (implements com.google.android.gms.maps.GoogleMap.OnCircleClickListener)</li>
-<li type="circle">fnsb.macstransit.ActivityListeners.<a href="../../../fnsb/macstransit/ActivityListeners/StopDeselected.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">StopDeselected</span></a> (implements com.google.android.gms.maps.GoogleMap.OnInfoWindowCloseListener)</li>
+<li type="circle">fnsb.macstransit.Activities.ActivityListeners.<a href="../../../fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">AdjustZoom</span></a> (implements com.google.android.gms.maps.GoogleMap.OnCameraIdleListener)</li>
+<li type="circle">fnsb.macstransit.Activities.ActivityListeners.<a href="../../../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">Helpers</span></a></li>
+<li type="circle">fnsb.macstransit.Activities.ActivityListeners.<a href="../../../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">InfoWindowAdapter</span></a> (implements com.google.android.gms.maps.GoogleMap.InfoWindowAdapter)</li>
+<li type="circle">fnsb.macstransit.Activities.ActivityListeners.<a href="../../../fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">StopClicked</span></a> (implements com.google.android.gms.maps.GoogleMap.OnCircleClickListener)</li>
+<li type="circle">fnsb.macstransit.Activities.ActivityListeners.<a href="../../../fnsb/macstransit/ActivityListeners/StopDeselected.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">StopDeselected</span></a> (implements com.google.android.gms.maps.GoogleMap.OnInfoWindowCloseListener)</li>
 </ul>
 </li>
 </ul>
