+++ docs/fnsb/macstransit/MapsActivity.html
@@ -120,7 +120,7 @@ var activeTableTab = "activeTableTab";
 <li>androidx.fragment.app.FragmentActivity</li>
 <li>
 <ul class="inheritance">
-<li>fnsb.macstransit.MapsActivity</li>
+<li>fnsb.macstransit.Activities.MapsActivity</li>
 </ul>
 </li>
 </ul>
