+++ docs/fnsb/macstransit/Network.html
@@ -99,7 +99,7 @@ var activeTableTab = "activeTableTab";
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>fnsb.macstransit.Network</li>
+<li>fnsb.macstransit.Threads.Network</li>
 </ul>
 </li>
 </ul>
