+++ docs/fnsb/macstransit/SplashActivity.html
@@ -123,7 +123,7 @@ var activeTableTab = "activeTableTab";
 <li>androidx.appcompat.app.AppCompatActivity</li>
 <li>
 <ul class="inheritance">
-<li>fnsb.macstransit.SplashActivity</li>
+<li>fnsb.macstransit.Activities.SplashActivity</li>
 </ul>
 </li>
 </ul>
