+++ docs/fnsb/macstransit/UpdateThread.html
@@ -93,7 +93,7 @@
 <li>java.lang.Object</li>
 <li>
 <ul class="inheritance">
-<li>fnsb.macstransit.UpdateThread</li>
+<li>fnsb.macstransit.Threads.UpdateThread</li>
 </ul>
 </li>
 </ul>
@@ -152,12 +152,12 @@ extends java.lang.Object</pre>
 <th class="colOne" scope="col">Constructor and Description</th>
 </tr>
 <tr class="altColor">
-<td class="colOne"><code><span class="memberNameLink"><a href="../../fnsb/macstransit/UpdateThread.html#UpdateThread-fnsb.macstransit.MapsActivity-">UpdateThread</a></span>(<a href="../../fnsb/macstransit/MapsActivity.html" title="class in fnsb.macstransit">MapsActivity</a>&nbsp;activity)</code>
+<td class="colOne"><code><span class="memberNameLink"><a href="../../fnsb/macstransit/UpdateThread.html#UpdateThread-fnsb.macstransit.Activities.MapsActivity-">UpdateThread</a></span>(<a href="../../fnsb/macstransit/MapsActivity.html" title="class in fnsb.macstransit">MapsActivity</a>&nbsp;activity)</code>
 <div class="block">Constructor for the UpdateThread.</div>
 </td>
 </tr>
 <tr class="rowColor">
-<td class="colOne"><code><span class="memberNameLink"><a href="../../fnsb/macstransit/UpdateThread.html#UpdateThread-fnsb.macstransit.MapsActivity-long-">UpdateThread</a></span>(<a href="../../fnsb/macstransit/MapsActivity.html" title="class in fnsb.macstransit">MapsActivity</a>&nbsp;activity,
+<td class="colOne"><code><span class="memberNameLink"><a href="../../fnsb/macstransit/UpdateThread.html#UpdateThread-fnsb.macstransit.Activities.MapsActivity-long-">UpdateThread</a></span>(<a href="../../fnsb/macstransit/MapsActivity.html" title="class in fnsb.macstransit">MapsActivity</a>&nbsp;activity,
             long&nbsp;updateFrequency)</code>
 <div class="block">Constructor for the UpdateThread.</div>
 </td>
@@ -210,7 +210,7 @@ extends java.lang.Object</pre>
 <!--   -->
 </a>
 <h3>Constructor Detail</h3>
-<a name="UpdateThread-fnsb.macstransit.MapsActivity-">
+<a name="UpdateThread-fnsb.macstransit.Activities.MapsActivity-">
 <!--   -->
 </a>
 <ul class="blockList">
@@ -224,7 +224,7 @@ extends java.lang.Object</pre>
 </dl>
 </li>
 </ul>
-<a name="UpdateThread-fnsb.macstransit.MapsActivity-long-">
+<a name="UpdateThread-fnsb.macstransit.Activities.MapsActivity-long-">
 <!--   -->
 </a>
 <ul class="blockListLast">
