+++ docs/index-files/index-1.html
@@ -73,25 +73,25 @@
 </a>
 <h2 class="title">A</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/Helpers.html#addCircle-com.google.android.gms.maps.GoogleMap-com.google.android.gms.maps.model.CircleOptions-java.lang.Object-boolean-">addCircle(GoogleMap, CircleOptions, Object, boolean)</a></span> - Static method in class fnsb.macstransit.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.ActivityListeners">Helpers</a></dt>
+<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/Helpers.html#addCircle-com.google.android.gms.maps.GoogleMap-com.google.android.gms.maps.model.CircleOptions-java.lang.Object-boolean-">addCircle(GoogleMap, CircleOptions, Object, boolean)</a></span> - Static method in class fnsb.macstransit.Activities.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.Activities.ActivityListeners">Helpers</a></dt>
 <dd>
 <div class="block">Helper function that adds a circle to the map with the provided options,
  and even sets the object this circle belongs to via its tag.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/Helpers.html#addMarker-com.google.android.gms.maps.GoogleMap-double-double-int-java.lang.String-java.lang.Object-">addMarker(GoogleMap, double, double, int, String, Object)</a></span> - Static method in class fnsb.macstransit.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.ActivityListeners">Helpers</a></dt>
+<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/Helpers.html#addMarker-com.google.android.gms.maps.GoogleMap-double-double-int-java.lang.String-java.lang.Object-">addMarker(GoogleMap, double, double, int, String, Object)</a></span> - Static method in class fnsb.macstransit.Activities.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.Activities.ActivityListeners">Helpers</a></dt>
 <dd>
 <div class="block">Helper function that adds a marker to the map at the provided location,
  with the provided title, and sets the color to the provided color.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/AdjustZoom.html#adjustCircleSize-float-fnsb.macstransit.RouteMatch.SharedStop:A-">adjustCircleSize(float, SharedStop[])</a></span> - Static method in class fnsb.macstransit.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.ActivityListeners">AdjustZoom</a></dt>
+<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/AdjustZoom.html#adjustCircleSize-float-fnsb.macstransit.RouteMatch.SharedStop:A-">adjustCircleSize(float, SharedStop[])</a></span> - Static method in class fnsb.macstransit.Activities.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.Activities.ActivityListeners">AdjustZoom</a></dt>
 <dd>
 <div class="block">Adjusts the circle size based on the current zoom level.</div>
 </dd>
-<dt><a href="../fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">AdjustZoom</span></a> - Class in <a href="../fnsb/macstransit/ActivityListeners/package-summary.html">fnsb.macstransit.ActivityListeners</a></dt>
+<dt><a href="../fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">AdjustZoom</span></a> - Class in <a href="../fnsb/macstransit/ActivityListeners/package-summary.html">fnsb.macstransit.Activities.ActivityListeners</a></dt>
 <dd>
 <div class="block">Created by Spud on 2019-10-28 for the project: MACS Transit.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/AdjustZoom.html#AdjustZoom-fnsb.macstransit.MapsActivity-">AdjustZoom(MapsActivity)</a></span> - Constructor for class fnsb.macstransit.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.ActivityListeners">AdjustZoom</a></dt>
+<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/AdjustZoom.html#AdjustZoom-fnsb.macstransit.Activities.MapsActivity-">AdjustZoom(MapsActivity)</a></span> - Constructor for class fnsb.macstransit.Activities.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.Activities.ActivityListeners">AdjustZoom</a></dt>
 <dd>
 <div class="block">Constructor for the listener.</div>
 </dd>
