+++ docs/index-files/index-12.html
@@ -73,12 +73,12 @@
 </a>
 <h2 class="title">O</h2>
 <dl>
-<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/AdjustZoom.html#onCameraIdle--">onCameraIdle()</a></span> - Method in class fnsb.macstransit.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.ActivityListeners">AdjustZoom</a></dt>
+<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/AdjustZoom.html#onCameraIdle--">onCameraIdle()</a></span> - Method in class fnsb.macstransit.Activities.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.Activities.ActivityListeners">AdjustZoom</a></dt>
 <dd>
 <div class="block">Called when camera movement has ended,
  there are no pending animations and the user has stopped interacting with the map.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/StopClicked.html#onCircleClick-com.google.android.gms.maps.model.Circle-">onCircleClick(Circle)</a></span> - Method in class fnsb.macstransit.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.ActivityListeners">StopClicked</a></dt>
+<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/StopClicked.html#onCircleClick-com.google.android.gms.maps.model.Circle-">onCircleClick(Circle)</a></span> - Method in class fnsb.macstransit.Activities.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.Activities.ActivityListeners">StopClicked</a></dt>
 <dd>
 <div class="block">Called when a circle is clicked.</div>
 </dd>
@@ -98,7 +98,7 @@
 <dd>
 <div class="block">Perform any final cleanup before an activity is destroyed.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/StopDeselected.html#onInfoWindowClose-com.google.android.gms.maps.model.Marker-">onInfoWindowClose(Marker)</a></span> - Method in class fnsb.macstransit.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/StopDeselected.html" title="class in fnsb.macstransit.ActivityListeners">StopDeselected</a></dt>
+<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/StopDeselected.html#onInfoWindowClose-com.google.android.gms.maps.model.Marker-">onInfoWindowClose(Marker)</a></span> - Method in class fnsb.macstransit.Activities.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/StopDeselected.html" title="class in fnsb.macstransit.Activities.ActivityListeners">StopDeselected</a></dt>
 <dd>
 <div class="block">Called when the marker's info window is closed.</div>
 </dd>
