+++ docs/index-files/index-15.html
@@ -132,19 +132,19 @@
 <dd>
 <div class="block">Loads the stop from the provided JSON.</div>
 </dd>
-<dt><a href="../fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">StopClicked</span></a> - Class in <a href="../fnsb/macstransit/ActivityListeners/package-summary.html">fnsb.macstransit.ActivityListeners</a></dt>
+<dt><a href="../fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">StopClicked</span></a> - Class in <a href="../fnsb/macstransit/ActivityListeners/package-summary.html">fnsb.macstransit.Activities.ActivityListeners</a></dt>
 <dd>
 <div class="block">Created by Spud on 2019-10-30 for the project: MACS Transit.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/StopClicked.html#StopClicked-fnsb.macstransit.MapsActivity-">StopClicked(MapsActivity)</a></span> - Constructor for class fnsb.macstransit.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.ActivityListeners">StopClicked</a></dt>
+<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/StopClicked.html#StopClicked-fnsb.macstransit.Activities.MapsActivity-">StopClicked(MapsActivity)</a></span> - Constructor for class fnsb.macstransit.Activities.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.Activities.ActivityListeners">StopClicked</a></dt>
 <dd>
 <div class="block">Constructor the the StopClicked listener.</div>
 </dd>
-<dt><a href="../fnsb/macstransit/ActivityListeners/StopDeselected.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">StopDeselected</span></a> - Class in <a href="../fnsb/macstransit/ActivityListeners/package-summary.html">fnsb.macstransit.ActivityListeners</a></dt>
+<dt><a href="../fnsb/macstransit/ActivityListeners/StopDeselected.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">StopDeselected</span></a> - Class in <a href="../fnsb/macstransit/ActivityListeners/package-summary.html">fnsb.macstransit.Activities.ActivityListeners</a></dt>
 <dd>
 <div class="block">Created by Spud on 2019-11-11 for the project: MACS Transit.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/StopDeselected.html#StopDeselected--">StopDeselected()</a></span> - Constructor for class fnsb.macstransit.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/StopDeselected.html" title="class in fnsb.macstransit.ActivityListeners">StopDeselected</a></dt>
+<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/StopDeselected.html#StopDeselected--">StopDeselected()</a></span> - Constructor for class fnsb.macstransit.Activities.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/StopDeselected.html" title="class in fnsb.macstransit.Activities.ActivityListeners">StopDeselected</a></dt>
 <dd>&nbsp;</dd>
 <dt><span class="memberNameLink"><a href="../fnsb/macstransit/RouteMatch/BasicStop.html#stopID">stopID</a></span> - Variable in class fnsb.macstransit.RouteMatch.<a href="../fnsb/macstransit/RouteMatch/BasicStop.html" title="class in fnsb.macstransit.RouteMatch">BasicStop</a></dt>
 <dd>
