+++ docs/index-files/index-16.html
@@ -81,11 +81,11 @@
 <dd>
 <div class="block">Created by Spud on 2019-10-13 for the project: MACS Transit.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../fnsb/macstransit/UpdateThread.html#UpdateThread-fnsb.macstransit.MapsActivity-">UpdateThread(MapsActivity)</a></span> - Constructor for class fnsb.macstransit.<a href="../fnsb/macstransit/UpdateThread.html" title="class in fnsb.macstransit">UpdateThread</a></dt>
+<dt><span class="memberNameLink"><a href="../fnsb/macstransit/UpdateThread.html#UpdateThread-fnsb.macstransit.Activities.MapsActivity-">UpdateThread(MapsActivity)</a></span> - Constructor for class fnsb.macstransit.<a href="../fnsb/macstransit/UpdateThread.html" title="class in fnsb.macstransit">UpdateThread</a></dt>
 <dd>
 <div class="block">Constructor for the UpdateThread.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../fnsb/macstransit/UpdateThread.html#UpdateThread-fnsb.macstransit.MapsActivity-long-">UpdateThread(MapsActivity, long)</a></span> - Constructor for class fnsb.macstransit.<a href="../fnsb/macstransit/UpdateThread.html" title="class in fnsb.macstransit">UpdateThread</a></dt>
+<dt><span class="memberNameLink"><a href="../fnsb/macstransit/UpdateThread.html#UpdateThread-fnsb.macstransit.Activities.MapsActivity-long-">UpdateThread(MapsActivity, long)</a></span> - Constructor for class fnsb.macstransit.<a href="../fnsb/macstransit/UpdateThread.html" title="class in fnsb.macstransit">UpdateThread</a></dt>
 <dd>
 <div class="block">Constructor for the UpdateThread.</div>
 </dd>
