+++ docs/index-files/index-5.html
@@ -79,7 +79,7 @@
 </dd>
 <dt><a href="../fnsb/macstransit/package-summary.html">fnsb.macstransit</a> - package fnsb.macstransit</dt>
 <dd>&nbsp;</dd>
-<dt><a href="../fnsb/macstransit/ActivityListeners/package-summary.html">fnsb.macstransit.ActivityListeners</a> - package fnsb.macstransit.ActivityListeners</dt>
+<dt><a href="../fnsb/macstransit/ActivityListeners/package-summary.html">fnsb.macstransit.Activities.ActivityListeners</a> - package fnsb.macstransit.Activities.ActivityListeners</dt>
 <dd>&nbsp;</dd>
 <dt><a href="../fnsb/macstransit/RouteMatch/package-summary.html">fnsb.macstransit.RouteMatch</a> - package fnsb.macstransit.RouteMatch</dt>
 <dd>&nbsp;</dd>
