+++ docs/index-files/index-6.html
@@ -89,11 +89,11 @@
 <dd>
 <div class="block">Gets the icon corresponding to the stop.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html#getInfoContents-com.google.android.gms.maps.model.Marker-">getInfoContents(Marker)</a></span> - Method in class fnsb.macstransit.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.ActivityListeners">InfoWindowAdapter</a></dt>
+<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html#getInfoContents-com.google.android.gms.maps.model.Marker-">getInfoContents(Marker)</a></span> - Method in class fnsb.macstransit.Activities.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.Activities.ActivityListeners">InfoWindowAdapter</a></dt>
 <dd>
 <div class="block">Provides custom contents for the default info window frame of a marker.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html#getInfoWindow-com.google.android.gms.maps.model.Marker-">getInfoWindow(Marker)</a></span> - Method in class fnsb.macstransit.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.ActivityListeners">InfoWindowAdapter</a></dt>
+<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html#getInfoWindow-com.google.android.gms.maps.model.Marker-">getInfoWindow(Marker)</a></span> - Method in class fnsb.macstransit.Activities.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.Activities.ActivityListeners">InfoWindowAdapter</a></dt>
 <dd>
 <div class="block">Provides a custom info window for a marker.</div>
 </dd>
@@ -113,7 +113,7 @@
 <dd>
 <div class="block">Gets the marker object for the stop.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/Helpers.html#getMarkerIcon-int-">getMarkerIcon(int)</a></span> - Static method in class fnsb.macstransit.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.ActivityListeners">Helpers</a></dt>
+<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/Helpers.html#getMarkerIcon-int-">getMarkerIcon(int)</a></span> - Static method in class fnsb.macstransit.Activities.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.Activities.ActivityListeners">Helpers</a></dt>
 <dd>
 <div class="block">Gets the color of the marker icon based off of the color value given.</div>
 </dd>
