+++ docs/index-files/index-8.html
@@ -77,11 +77,11 @@
 <dd>
 <div class="block">The options that correspond to the stop icon (which is a circle).</div>
 </dd>
-<dt><a href="../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">InfoWindowAdapter</span></a> - Class in <a href="../fnsb/macstransit/ActivityListeners/package-summary.html">fnsb.macstransit.ActivityListeners</a></dt>
+<dt><a href="../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">InfoWindowAdapter</span></a> - Class in <a href="../fnsb/macstransit/ActivityListeners/package-summary.html">fnsb.macstransit.Activities.ActivityListeners</a></dt>
 <dd>
 <div class="block">Created by Spud on 2019-11-01 for the project: MACS Transit.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html#InfoWindowAdapter-fnsb.macstransit.MapsActivity-">InfoWindowAdapter(MapsActivity)</a></span> - Constructor for class fnsb.macstransit.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.ActivityListeners">InfoWindowAdapter</a></dt>
+<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html#InfoWindowAdapter-fnsb.macstransit.Activities.MapsActivity-">InfoWindowAdapter(MapsActivity)</a></span> - Constructor for class fnsb.macstransit.Activities.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.Activities.ActivityListeners">InfoWindowAdapter</a></dt>
 <dd>
 <div class="block">Constructor for the info window adapter.</div>
 </dd>
