+++ docs/index-files/index-9.html
@@ -85,7 +85,7 @@
 <dd>
 <div class="block">The latitude and longitude of the SharedStops.</div>
 </dd>
-<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/Helpers.html#loadAllStops-fnsb.macstransit.RouteMatch.Route:A-">loadAllStops(Route[])</a></span> - Static method in class fnsb.macstransit.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.ActivityListeners">Helpers</a></dt>
+<dt><span class="memberNameLink"><a href="../fnsb/macstransit/ActivityListeners/Helpers.html#loadAllStops-fnsb.macstransit.RouteMatch.Route:A-">loadAllStops(Route[])</a></span> - Static method in class fnsb.macstransit.Activities.ActivityListeners.<a href="../fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.Activities.ActivityListeners">Helpers</a></dt>
 <dd>
 <div class="block">Loads all the stops from the provided routes.</div>
 </dd>
