+++ docs/overview-frame.html
@@ -14,7 +14,7 @@
 <h2 title="Packages">Packages</h2>
 <ul title="Packages">
 <li><a href="fnsb/macstransit/package-frame.html" target="packageFrame">fnsb.macstransit</a></li>
-<li><a href="fnsb/macstransit/ActivityListeners/package-frame.html" target="packageFrame">fnsb.macstransit.ActivityListeners</a></li>
+<li><a href="fnsb/macstransit/ActivityListeners/package-frame.html" target="packageFrame">fnsb.macstransit.Activities.ActivityListeners</a></li>
 <li><a href="fnsb/macstransit/RouteMatch/package-frame.html" target="packageFrame">fnsb.macstransit.RouteMatch</a></li>
 </ul>
 </div>
