+++ docs/overview-summary.html
@@ -81,7 +81,7 @@
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr class="rowColor">
-<td class="colFirst"><a href="fnsb/macstransit/ActivityListeners/package-summary.html">fnsb.macstransit.ActivityListeners</a></td>
+<td class="colFirst"><a href="fnsb/macstransit/ActivityListeners/package-summary.html">fnsb.macstransit.Activities.ActivityListeners</a></td>
 <td class="colLast">&nbsp;</td>
 </tr>
 <tr class="altColor">
