+++ docs/overview-tree.html
@@ -73,7 +73,7 @@
 <span class="packageHierarchyLabel">Package Hierarchies:</span>
 <ul class="horizontal">
 <li><a href="fnsb/macstransit/package-tree.html">fnsb.macstransit</a>, </li>
-<li><a href="fnsb/macstransit/ActivityListeners/package-tree.html">fnsb.macstransit.ActivityListeners</a>, </li>
+<li><a href="fnsb/macstransit/ActivityListeners/package-tree.html">fnsb.macstransit.Activities.ActivityListeners</a>, </li>
 <li><a href="fnsb/macstransit/RouteMatch/package-tree.html">fnsb.macstransit.RouteMatch</a></li>
 </ul>
 </div>
@@ -82,7 +82,7 @@
 <ul>
 <li type="circle">java.lang.Object
 <ul>
-<li type="circle">fnsb.macstransit.ActivityListeners.<a href="fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">AdjustZoom</span></a> (implements com.google.android.gms.maps.GoogleMap.OnCameraIdleListener)</li>
+<li type="circle">fnsb.macstransit.Activities.ActivityListeners.<a href="fnsb/macstransit/ActivityListeners/AdjustZoom.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">AdjustZoom</span></a> (implements com.google.android.gms.maps.GoogleMap.OnCameraIdleListener)</li>
 <li type="circle">fnsb.macstransit.RouteMatch.<a href="fnsb/macstransit/RouteMatch/BasicStop.html" title="class in fnsb.macstransit.RouteMatch"><span class="typeNameLink">BasicStop</span></a>
 <ul>
 <li type="circle">fnsb.macstransit.RouteMatch.<a href="fnsb/macstransit/RouteMatch/Stop.html" title="class in fnsb.macstransit.RouteMatch"><span class="typeNameLink">Stop</span></a></li>
@@ -123,14 +123,14 @@
 </li>
 </ul>
 </li>
-<li type="circle">fnsb.macstransit.ActivityListeners.<a href="fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">Helpers</span></a></li>
-<li type="circle">fnsb.macstransit.ActivityListeners.<a href="fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">InfoWindowAdapter</span></a> (implements com.google.android.gms.maps.GoogleMap.InfoWindowAdapter)</li>
+<li type="circle">fnsb.macstransit.Activities.ActivityListeners.<a href="fnsb/macstransit/ActivityListeners/Helpers.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">Helpers</span></a></li>
+<li type="circle">fnsb.macstransit.Activities.ActivityListeners.<a href="fnsb/macstransit/ActivityListeners/InfoWindowAdapter.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">InfoWindowAdapter</span></a> (implements com.google.android.gms.maps.GoogleMap.InfoWindowAdapter)</li>
 <li type="circle">fnsb.macstransit.<a href="fnsb/macstransit/Network.html" title="class in fnsb.macstransit"><span class="typeNameLink">Network</span></a></li>
 <li type="circle">fnsb.macstransit.RouteMatch.<a href="fnsb/macstransit/RouteMatch/Route.html" title="class in fnsb.macstransit.RouteMatch"><span class="typeNameLink">Route</span></a></li>
 <li type="circle">fnsb.macstransit.RouteMatch.<a href="fnsb/macstransit/RouteMatch/RouteMatch.html" title="class in fnsb.macstransit.RouteMatch"><span class="typeNameLink">RouteMatch</span></a></li>
 <li type="circle">fnsb.macstransit.RouteMatch.<a href="fnsb/macstransit/RouteMatch/SharedStop.html" title="class in fnsb.macstransit.RouteMatch"><span class="typeNameLink">SharedStop</span></a></li>
-<li type="circle">fnsb.macstransit.ActivityListeners.<a href="fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">StopClicked</span></a> (implements com.google.android.gms.maps.GoogleMap.OnCircleClickListener)</li>
-<li type="circle">fnsb.macstransit.ActivityListeners.<a href="fnsb/macstransit/ActivityListeners/StopDeselected.html" title="class in fnsb.macstransit.ActivityListeners"><span class="typeNameLink">StopDeselected</span></a> (implements com.google.android.gms.maps.GoogleMap.OnInfoWindowCloseListener)</li>
+<li type="circle">fnsb.macstransit.Activities.ActivityListeners.<a href="fnsb/macstransit/ActivityListeners/StopClicked.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">StopClicked</span></a> (implements com.google.android.gms.maps.GoogleMap.OnCircleClickListener)</li>
+<li type="circle">fnsb.macstransit.Activities.ActivityListeners.<a href="fnsb/macstransit/ActivityListeners/StopDeselected.html" title="class in fnsb.macstransit.Activities.ActivityListeners"><span class="typeNameLink">StopDeselected</span></a> (implements com.google.android.gms.maps.GoogleMap.OnInfoWindowCloseListener)</li>
 <li type="circle">fnsb.macstransit.<a href="fnsb/macstransit/UpdateThread.html" title="class in fnsb.macstransit"><span class="typeNameLink">UpdateThread</span></a></li>
 </ul>
 </li>
