+++ docs/package-list
@@ -1,3 +1,3 @@
 fnsb.macstransit
-fnsb.macstransit.ActivityListeners
+fnsb.macstransit.Activities.ActivityListeners
 fnsb.macstransit.RouteMatch
